requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
email-validator==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Reuploaded==0.3.2
Flask-WTF==0.14.3
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
requests==2.24.0
Pillow==8.0.1
shopyo
pymysql
cryptography
pandas
